+++ bfv/ciphertext.go
@@ -9,12 +9,12 @@ type Ciphertext struct {
 
 // NewCiphertext creates a new ciphertext parameterized by degree, level and scale.
 func NewCiphertext(params *Parameters, degree uint64) (ciphertext *Ciphertext) {
-	return &Ciphertext{newEleCT(params, degree)}
+	return &Ciphertext{newCiphertextElement(params, degree)}
 }
 
 // NewCiphertextRandom generates a new uniformly distributed ciphertext of degree, level and scale.
 func NewCiphertextRandom(prng utils.PRNG, params *Parameters, degree uint64) (ciphertext *Ciphertext) {
-	ciphertext = &Ciphertext{newEleCT(params, degree)}
+	ciphertext = &Ciphertext{newCiphertextElement(params, degree)}
 	populateElementRandom(prng, params, ciphertext.Element)
 	return
 }