+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 ## Unreleased changes
 
 ### Added
-- BFV : special-purpose plaintext types (NewPlaintextZQ, NewPlaintextZT or NewPlaintextMul) optimized for different operations. See bfv/plaintext.go.
-- RING : GenNTTPrimes now takes the Nth primitive root as input rather than logN.
+- BFV : special-purpose plaintext types (`PlaintextRingT` or `PlaintextMul`) optimized ct-pt operations. See bfv/plaintext.go.
+- BFV : allocation-free `Encoder` methods
+- RING : `GenNTTPrimes` now takes the Nth primitive root as input rather than logN.
 
 ### Changed
+- BFV : the `Encoder.DecodeUint64` and `Encoder.DecodeInt64` methods now take the output slice as argument.
 - CKKS : API of evaluator.RotateColumns becomes Evaluator.Rotate
 - CKKS : the change of variable in evaluator.EvaluateCheby isn't done automatically anymore and the user must do it before calling the function to ensure correctness.
 - CKKS : when encoding, the number of slots must now be given in log2 basis. This is to prevent errors that would induced by zero values or non power of two values.